2010-01-03  Jean Brefort  <jean.brefort@normalesup.org> 	* libs/gcp/document.cc: improve portability. 	* libs/gcp/fragment-atom.cc: fixed compilation warning. 	* libs/gcp/fragment.cc: ditto. 	* plugins/loaders/cdxml/cdxml.cc: ditto. 	* plugins/paint/wikipedia/wikipediatool.cc: improve portability. 	* programs/spectra/application.cc: fixed compilation warning. 	* programs/table/gchemtable-app.cc: ditto.

+++ libs/gcp/fragment.cc
@@ -1042,7 +1042,7 @@ Object* Fragment::GetAtomAt (double x, double y, double z)
 		cur--;
 		c = m_buf[cur];
 	}
-	if (index - cur > Residue::MaxSymbolLength)
+	if (index - cur > static_cast <int> (Residue::MaxSymbolLength))
 		cur = index - Residue::MaxSymbolLength;
 	if (cur < 0)
 		cur = 0;
